argmax_mini/requirements.txt
adrf==0.1.9
alembic==1.14.1
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
async-property==0.2.2
bayesian-optimization==2.0.3
catboost==1.2.7
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
deap==1.4.2
Django==4.2.18
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
einops==0.8.0
faiss-gpu==1.7.2
filelock==3.17.0
fonttools==4.55.8
fsspec==2025.2.0
graphviz==0.20.3
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflection==0.5.1
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
lightgbm==4.5.0
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
narwhals==1.24.2
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
optuna==4.2.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
polars==1.21.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
sympy==1.13.1
tabpfn==2.0.5
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
zipp==3.21.0
pyarrow==19.0.0
fireducks==1.2.1